'sdfsadf'
Former-commit-id: 6e63759ee9791b00a9ac8298f337d6261ff72e7a [formerly 17c7ae81f627e3a5438dc65dbf72ab022005b8b3]
Former-commit-id: 50320a192202cf5c8c88ae234065205514deb026

+++ Data/FuncionesGenerales.cs
@@ -2737,7 +2737,7 @@ namespace CerealNet.WSCartasDePorte
         //public bool brutodestSpecified;
         public string CPoriginal;
         public string calidad;
-        public int cartaporte;
+        public long cartaporte;
         public int codmerca;
         //public bool codmercaSpecified;
         public int codonccalocalidadpuerto;
@@ -2793,7 +2793,7 @@ namespace CerealNet.WSCartasDePorte
 
 
         // version 2
-        public int CEE;
+        public long CEE;
         public System.DateTime fechaEmisionCarga;
         public System.DateTime fechavencimiento;
         public int CTG;
